+++ livechat/customer/rtm/api/v36.py
@@ -24,24 +24,27 @@ class CustomerRtmV36:
 
     def open_connection(self,
                         origin: dict = None,
-                        ping_timeout: float = 3,
+                        ping_timeout: float = 30,
                         ping_interval: float = 5,
                         ws_conn_timeout: float = 10,
-                        keep_alive: bool = True) -> None:
+                        keep_alive: bool = True,
+                        response_timeout: float = 3) -> None:
         ''' Opens WebSocket connection.
 
             Args:
                 origin (dict): Specifies origin while creating websocket connection.
                 ping_timeout (int or float): timeout (in seconds) if the pong message is not received,
-                    by default sets to 3 seconds.
+                    by default sets to 30 seconds.
                 ping_interval (int or float): automatically sends "ping" command every specified period (in seconds).
                     If set to 0, no ping is sent periodically, by default sets to 5 seconds.
                 ws_conn_timeout (int or float): timeout (in seconds) to wait for WebSocket connection,
                     by default sets to 10 seconds.
                 keep_alive(bool): Bool which states if connection should be kept, by default sets to `True`.
+                response_timeout (int or float): timeout (in seconds) to wait for the response,
+                    by default sets to 3 seconds.
         '''
         self.ws.open(origin, ping_timeout, ping_interval, ws_conn_timeout,
-                     keep_alive)
+                     keep_alive, response_timeout)
 
     def close_connection(self) -> None:
         ''' Closes WebSocket connection. '''