+++ livechat/utils/ws_client.py
@@ -7,7 +7,7 @@ import random
 import ssl
 import threading
 from time import sleep
-from typing import List, NoReturn
+from typing import List, NoReturn, Union
 
 from loguru import logger
 from websocket import WebSocketApp, WebSocketConnectionClosedException
@@ -27,13 +27,15 @@ class WebsocketClient(WebSocketApp):
         super().__init__(*args, **kwargs)
         self.messages: List[dict] = []
         self.on_message = on_message
+        self.response_timeout = None
 
     def open(self,
              origin: dict = None,
-             ping_timeout: float = 3,
-             ping_interval: float = 5,
-             ws_conn_timeout: float = 10,
-             keep_alive: bool = True) -> NoReturn:
+             ping_timeout: Union[float, int] = 3,
+             ping_interval: Union[float, int] = 5,
+             ws_conn_timeout: Union[float, int] = 10,
+             keep_alive: bool = True,
+             response_timeout: Union[float, int] = 3) -> NoReturn:
         ''' Opens websocket connection and keep running forever.
             Args:
                 origin (dict): Specifies origin while creating websocket connection.
@@ -43,7 +45,10 @@ class WebsocketClient(WebSocketApp):
                     If set to 0, no ping is sent periodically, by default sets to 5 seconds.
                 ws_conn_timeout (int or float): timeout (in seconds) to wait for WebSocket connection,
                     by default sets to 10 seconds.
-                keep_alive(bool): Bool which states if connection should be kept, by default sets to `True`. '''
+                keep_alive(bool): Bool which states if connection should be kept, by default sets to `True`.
+                response_timeout (int or float): timeout (in seconds) to wait for the response,
+                    by default sets to 3 seconds. '''
+        self.response_timeout = response_timeout
         run_forever_kwargs = {
             'sslopt': {
                 'cert_reqs': ssl.CERT_NONE
@@ -60,17 +65,13 @@ class WebsocketClient(WebSocketApp):
             return
         self.run_forever(**run_forever_kwargs)
 
-    def send(self,
-             request: dict,
-             opcode=ABNF.OPCODE_TEXT,
-             response_timeout=2) -> dict:
+    def send(self, request: dict, opcode=ABNF.OPCODE_TEXT) -> dict:
         '''
         Sends message, assigining a random request ID, fetching and returning response(s).
             Args:
                 request (dict): message to send. If you set opcode to OPCODE_TEXT,
                     data must be utf-8 string or unicode.
                 opcode (int): operation code of data. default is OPCODE_TEXT.
-                response_timeout (int): time in seconds to wait for the response.
 
             Returns:
                 RtmResponse: RTM response structure (`request_id`, `action`,
@@ -87,13 +88,14 @@ class WebsocketClient(WebSocketApp):
             (item
              for item in self.messages if item.get('request_id') == request_id
              and item.get('type') == 'response'),
-                None)) and response_timeout > 0:
+                None)) and self.response_timeout > 0:
             sleep(0.2)
-            response_timeout -= 0.2
+            self.response_timeout -= 0.2
         logger.info(f'\nRESPONSE:\n{json.dumps(response, indent=4)}')
         return RtmResponse(response)
 
-    def _wait_till_sock_connected(self, timeout: float = 10) -> NoReturn:
+    def _wait_till_sock_connected(self,
+                                  timeout: Union[float, int] = 10) -> NoReturn:
         ''' Polls until `self.sock` is connected.
             Args:
                 timeout (float): timeout value in seconds, default 10. '''