+++ changelog.md
@@ -6,6 +6,9 @@ All notable changes to this project will be documented in this file.
 ### Added
 - New `get_company_details` method in configuration-api v3.6.
 
+### Changed
+- Updated outdated packages.
+
 ## [0.3.9] - 2024-04-22
 
 ### Added