+++ changelog.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 - Updated outdated packages.
+- Exposed `response_timeout` parameter in `open_connection` methods.
+- Changed default value of `ping_timeout` to 30 seconds.
 
 ## [0.3.9] - 2024-04-22